requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.2.0
bcrypt==4.1.2
certifi==2024.12.14
click==8.1.7
dnspython==2.7.0
ecdsa==0.18.0
email_validator==2.2.0
fastapi==0.115.6
fastapi-cli==0.0.7
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
idna==3.6
Jinja2==3.1.3
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.5.1
pydantic==2.6.0
pydantic-settings==2.1.0
pydantic_core==2.16.1
Pygments==2.18.0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.7
PyYAML==6.0.2
rich==13.9.4
rich-toolkit==0.12.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sqladmin==0.16.0
SQLAlchemy==2.0.25
starlette==0.41.3
typer==0.15.1
typing_extensions==4.12.2
uvicorn==0.27.0.post1
uvloop==0.21.0
watchfiles==1.0.3
websockets==14.1
WTForms==3.1.2
litellm
boto3